Fix #33: fix HankkiTopBar to use more free
